Markups - get S4 to call SM. All UT passed
