+++ src/Makefile
@@ -207,7 +207,8 @@ sprout_test_SOURCES := ${SPROUT_COMMON_SOURCES} \
                        rphservice_test.cpp \
                        mock_rph_service.cpp \
                        s4_test.cpp \
-                       mock_s4.cpp
+                       mock_s4.cpp \
+                       s4_chronoshandlers_test.cpp
 
 COVERAGE_ROOT := ..
 sprout_test_COVERAGE_EXCLUSIONS := ^src/ut|^usr|^modules/gmock|^modules/cpp-common|^modules/rapidjson|^include|^src/mangelwurzel/ut|^modules/gemini/src/ut|^modules/gemini/include|^modules/clearwater-s4/src/ut
