+++ src/ut/chronoshandlers_test.cpp
@@ -28,98 +28,7 @@ using namespace std;
 using namespace TestingCommon;
 using ::testing::_;
 using ::testing::Return;
-using ::testing::SetArgReferee;
-using ::testing::SaveArg;
 using ::testing::InSequence;
-using ::testing::NiceMock;
-
-// Test that S4 handles request from Chronos about AoRTimeout, by sending back
-// response and calling into SM.
-class ChronosAoRTimeoutTasksTest : public TestWithMockSdms
-{
-public:
-  void TearDown()
-  {
-    delete config;
-    delete req;
-
-    TestWithMockSdms::TearDown();
-  }
-
-  void build_timeout_request(std::string body, htp_method method)
-  {
-    req = new MockHttpStack::Request(stack, "/", "timers", "", body, method);
-    config = new AoRTimeoutTask::Config(sm);
-    handler = new ChronosAoRTimeoutTask(*req, config, 0);
-  }
-
-  MockHttpStack::Request* req;
-  AoRTimeoutTask::Config* config;
-  ChronosAoRTimeoutTask* handler;
-};
-
-// Mainline test where the request is successfully parsed and handler proceeds
-// to call SM. 
-TEST_F(ChronosAoRTimeoutTasksTest, MainlineTest)
-{
-  // Build request from Chronos
-  std::string aor_id = "sip:6505550231@homedomain";
-  std::string body = "{\"aor_id\": \"sip:6505550231@homedomain\"}";
-  build_timeout_request(body, htp_method_POST);
-
-  {
-    InSequence s;
-      // Send back response as soon as the request is successfully parsed
-      EXPECT_CALL(*stack, send_reply(_, 200, _));
-      EXPECT_CALL(*sm, handle_timer_pop(_, _));
-  }
-  handler->run();
-}
-
-
-/// The following tests deal with error cases in parsing the timer pop request.
-
-// Test that an invalid HTTP method fails with HTTP_BADMETHOD
-TEST_F(ChronosAoRTimeoutTasksTest, InvalidHTTPMethodTest)
-{
-  std::string body = "{\"aor_id\": \"sip:6505550231@homedomain\"}";
-  build_timeout_request(body, htp_method_PUT);
-
-  EXPECT_CALL(*stack, send_reply(_, 405, _));
-
-  handler->run();
-}
-
-// Test that an invalid JSON body fails in parsing
-TEST_F(ChronosAoRTimeoutTasksTest, InvalidJSONTest)
-{
-  CapturingTestLogger log(5);
-
-  std::string body = "{\"aor_id\" \"aor_id\"}";
-  build_timeout_request(body, htp_method_POST);
-
-  EXPECT_CALL(*stack, send_reply(_, 400, _));
-
-  handler->run();
-
-  EXPECT_TRUE(log.contains("Failed to parse opaque data as JSON:"));
-}
-
-// Test that a body without an AoR ID fails, logging "Badly formed opaque data"
-TEST_F(ChronosAoRTimeoutTasksTest, MissingAorJSONTest)
-{
-  CapturingTestLogger log(5);
-
-  std::string body = "{}";
-  build_timeout_request(body, htp_method_POST);
-
-  EXPECT_CALL(*stack, send_reply(_, 400, _));
-
-  handler->run();
-
-  EXPECT_TRUE(log.contains("Badly formed opaque data (missing aor_id)"));
-}
-
 
 class ChronosAuthTimeoutTest : public AuthTimeoutTest
 {
